requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
py-clob-client>=0.34.6
python-dotenv>=1.0.0
